neglogpost: add comment; TODO
NEEDS REVISION
when xhat is saved to par.fxhat in neglogpost, xhat includes the structure allparams. this stucture does not exist when xhat is saved inside the PrintPar function.
It would be better to not write and overwrite the same xhat file every iteration, so need to pick which location makes more sense to execute the save and then make sure all the necessary fields are included there.
- for ResetPar, x0 has to be a variable saved in par.fxhat. ResetPar is run in each of the first 10 iterations of optimization.
- for SetPar, xhat structure should include a substructure called allparams. (SetPar is run once per driver run)

TODO: move save allparams to inside PrintPar. Then, remove the save command for xhat in neglogpost.