(LOOK) eqPcycle: add | ~any() to allow for optimizing C or O model parameters without turning on P model parameters
useful for GHtest, if you want to test just C or O model parameter gradients without needing any P model parameters to be on.
Without this, the model requires at least one parameter in each of the element cycle models (Pmodel, Cmodel, and Omodel) to be on.

I initially put this in to code so I could use a common fixed P cycle model and optimize only carbon cycle parameters with different C2P parameterizations